+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 ### Feat
 
+- **LactoLine**: create Priority metric
+- **LactoLine**: create Product class
+- **LactoLine**: add Date module
 - **LactoLine**: add Progress object
 - **LactoLine**: add Amont object
 - **LactoLine**: create Name object
 - **LactoLine**: add enviroment files
+
+### Fix
+
+- **LactoLine**: fix Date class
+
+### Refactor
+
+- **LactoLine**: swap days and years params at DaysDistance